+++ internal/jobqueue/jobqueue.go
@@ -2263,9 +2263,10 @@ func (w *WebhookRemovalWorker) updateWebhookRegistryForGiteaRemoval(ctx context.
 
 // JobQueue manages the River job queue
 type JobQueue struct {
-	client *river.Client[pgx.Tx]
-	pool   *pgxpool.Pool
-	config *QueueConfig
+	client                *river.Client[pgx.Tx]
+	pool                  *pgxpool.Pool
+	db                    *sql.DB
+	config                *QueueConfig
 }
 
 // NewJobQueue creates a new job queue instance
@@ -2291,9 +2292,13 @@ func NewJobQueue(databaseURL string, db *sql.DB) (*JobQueue, error) {
 	}
 	httpClient := networkjobqueue.NewWebhookHTTPClient(30 * time.Second)
 
+	webhookWorker := &WebhookReviewWorker{}
+	manualWorker := &ManualReviewWorker{}
 	river.AddWorker(workers, &WebhookInstallWorker{pool: pool, config: config, store: store, httpClient: httpClient})
 	river.AddWorker(workers, &WebhookRemovalWorker{pool: pool, config: config, store: store, httpClient: httpClient})
 	river.AddWorker(workers, &DiffReviewWorker{db: db, pool: pool})
+	river.AddWorker(workers, webhookWorker)
+	river.AddWorker(workers, manualWorker)
 
 	client, err := river.NewClient(riverpgxv5.New(pool), &river.Config{
 		Queues:  config.RiverQueueConfig(),
@@ -2303,12 +2308,18 @@ func NewJobQueue(databaseURL string, db *sql.DB) (*JobQueue, error) {
 		return nil, fmt.Errorf("failed to create River client: %w", err)
 	}
 
-	return &JobQueue{
+	jq := &JobQueue{
 		client: client,
 		pool:   pool,
+		db:     db,
 		config: config,
-	}, nil
+	}
+	webhookWorker.jq = jq
+	manualWorker.jq = jq
+
+	return jq, nil
 }
+
 
 // Start starts the job queue workers
 func (jq *JobQueue) Start(ctx context.Context) error {
@@ -2362,6 +2373,38 @@ func (jq *JobQueue) QueueReviewJob(ctx context.Context, args DiffReviewJobArgs) 
 	_, err := jq.client.Insert(ctx, args, &river.InsertOpts{Queue: "review"})
 	if err != nil {
 		return fmt.Errorf("failed to queue review job: %w", err)
+	}
+	return nil
+}
+
+// QueueWebhookReviewJob enqueues a new webhook review job to the "review" queue.
+func (jq *JobQueue) QueueWebhookReviewJob(ctx context.Context, orgID int64, connectorID int64, eventJSON string, scenarioType string) error {
+	args := WebhookReviewJobArgs{
+		OrgID:        orgID,
+		ConnectorID:  connectorID,
+		EventJSON:    eventJSON,
+		ScenarioType: scenarioType,
+	}
+	_, err := jq.client.Insert(ctx, args, &river.InsertOpts{Queue: "review"})
+	if err != nil {
+		return fmt.Errorf("failed to queue webhook review job: %w", err)
+	}
+	return nil
+}
+
+// QueueManualReviewJob enqueues a new manual review job to the "review" queue.
+func (jq *JobQueue) QueueManualReviewJob(ctx context.Context, orgID int64, planCode string, actorUserID *int64, actorEmail string, reviewID int64, requestJSON string) error {
+	args := ManualReviewJobArgs{
+		OrgID:       orgID,
+		PlanCode:    planCode,
+		ActorUserID: actorUserID,
+		ActorEmail:  actorEmail,
+		ReviewID:    reviewID,
+		RequestJSON: requestJSON,
+	}
+	_, err := jq.client.Insert(ctx, args, &river.InsertOpts{Queue: "review"})
+	if err != nil {
+		return fmt.Errorf("failed to queue manual review job: %w", err)
 	}
 	return nil
 }
